improved xas spectra analysis
changes to spectra_script.py, xas_notebook.ipynb to use new Spectra objects.

Old functions removed.

Various other improvements.
