+++ environment.yml
@@ -24,4 +24,4 @@ dependencies:
       - hdfmap
       # and for wheels
       #- http://www.lfd.uci.edu/~gohlke/pythonlibs/bofhrmxk/opencv_python-3.1.0-cp35-none-win_amd64.whl
-prefix: /dls/science/users/grp66007/envs/mmg_toolbox
+# prefix: /dls/science/users/grp66007/envs/mmg_toolbox
