+++ mmg_toolbox/env_functions.py
@@ -8,13 +8,13 @@ import subprocess
 import tempfile
 from datetime import datetime
 
+from mmg_toolbox.file_functions import get_scan_number
+
 # environment variables on beamline computers
 BEAMLINE = 'BEAMLINE'
 USER = ['USER', 'USERNAME']
 DLS = '/dls'
 MMG_BEAMLINES = ['i06', 'i06-1', 'i06-2', 'i10', 'i10-1', 'i16', 'i21']
-
-regex_scan_number = re.compile(r'\d{3,}')
 
 # Find writable directory
 TMPDIR = tempfile.gettempdir()
@@ -77,22 +77,6 @@ def get_dls_visits(instrument: str | None = None, year: str | int | None = None)
     if os.path.isdir(dls_dir):
         return {p.name: p.path for p in os.scandir(dls_dir) if p.is_dir()}
     return {}
-
-
-def get_scan_number(filename: str) -> int:
-    """Return scan number from scan filename"""
-    filename = os.path.basename(filename)
-    match = regex_scan_number.search(filename)
-    if match:
-        return int(match[0])
-    return 0
-
-
-def replace_scan_number(filename: str, new_number: int) -> str:
-    """Replace scan number in filename"""
-    path, filename = os.path.split(filename)
-    new_filename = regex_scan_number.sub(str(new_number), filename)
-    return os.path.join(path, new_filename)
 
 
 def get_first_file(folder: str, extension='.nxs') -> str:
