+++ mmg_toolbox/nexus_reader.py
@@ -1,5 +1,5 @@
 
-import os
+
 import hdfmap
 
 from mmg_toolbox.misc_functions import DataHolder, data_holder
@@ -26,26 +26,3 @@ def read_nexus_file(filename: str) -> DataHolder:
     hdf_map = hdfmap.create_nexus_map(filename)
     return nexus_data_holder(hdf_map)
 
-
-"""
-Old functions below - remove!
-"""
-
-def read_Nexus_file(filename: str):
-    return hdfmap.nexus_data_block(filename)
-
-
-def readscan_Nexus(num):
-    if os.path.isdir(filedir) == False:
-        print("I can't find the directory: {}".format(filedir))
-        return None
-
-    file = os.path.join(filedir, nxsfile_format % num)
-
-    try:
-        d = read_Nexus_file(file)
-        # d = dnp.io.load(file,warn=False) # from SciSoftPi
-    except:
-        print("Scan {} doesn't exist or can't be read".format(num))
-        return None
-    return (d)
