+++ mmg_toolbox/tkguis/widgets/script_runner.py
@@ -7,8 +7,8 @@ import tkinter as tk
 from tkinter import ttk
 
 from ...scripts import scripts
-from ...env_functions import (get_scan_numbers, get_last_scan_number, get_first_file,
-                              replace_scan_number, get_scan_number)
+from ...env_functions import (get_scan_numbers, get_last_scan_number, get_first_file)
+from ...file_functions import get_scan_number, replace_scan_number
 from ..misc.logging import create_logger
 from ..misc.config import get_config
 from ..misc.functions import select_folder
